Improve the APIs
1) Improve the documentation of two extension points (Lifecycle Manager
Provider and Tab Descriptor Provider)
2) Remove two useless extension points along with their code (View
Description Provider and Interpreter Provider)
3) Move some interfaces related to Lifecycle Managers in order to
have all APIs for the Lifecycle Managers (2 abstract classes and 2
interfaces) in the same package.
4) Move the two classes of o.e.eef.properties.ui.legacy in an internal
package since they are only used in this bundle (I18N & Messages)
5) Update the behavior of the Lifecycle Manager Provider extension point
to let contibutors replace existing Lifecycle Managers

+++ plugins/org.eclipse.eef.ide.ui/src/org/eclipse/eef/ide/ui/api/widgets/AbstractEEFWidgetLifecycleManager.java
@@ -348,7 +348,7 @@ public abstract class AbstractEEFWidgetLifecycleManager extends AbstractEEFLifec
 	/**
 	 * {@inheritDoc}
 	 *
-	 * @see org.eclipse.eef.ide.ui.api.ILifecycleManager#dispose()
+	 * @see org.eclipse.eef.ide.ui.api.widgets.IEEFLifecycleManager#dispose()
 	 */
 	@Override
 	public void dispose() {
@@ -488,7 +488,7 @@ public abstract class AbstractEEFWidgetLifecycleManager extends AbstractEEFLifec
 			Eval eval = new Eval(interpreter, variableManager);
 			Integer fontSizeValue = eval.get(eAttribute, fontSizeExpression, Integer.class);
 			if (fontSizeValue != null && fontSizeValue.intValue() != fontSize) {
-				fontSize = fontSizeValue;
+				fontSize = fontSizeValue.intValue();
 			}
 		}
 		return fontSize;
